requirements.txt
requests
opencv-python
numpy
Pillow
streamlit
streamlit-image-coordinates
streamlit-dimensions